# Enable readability-container-contains clang-tidy check