closes https://github.com/Project-OSRM/osrm-backend/issues/6902

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1085.98<br/>plain u32: 1090.84<br/>aliased double: 949.921<br/>plain double: 946.638 | aliased u32: 1085.29<br/>plain u32: 1078.87<br/>aliased double: 940.223<br/>plain double: 946.212 |
| json-render | String: 6.81824ms<br/>Stringstream: 10.6523ms<br/>Vector: 6.62559ms | String: 6.78835ms<br/>Stringstream: 10.5756ms<br/>Vector: 6.60494ms |
| match_ch | Default radius: <br/>4.5085ms/req at 82 coordinate<br/>0.0549817ms/coordinate<br/>Radius 5m: <br/>4.49103ms/req at 82 coordinate<br/>0.0547686ms/coordinate<br/>Radius 10m: <br/>15.2236ms/req at 82 coordinate<br/>0.185654ms/coordinate<br/>Radius 15m: <br/>37.0803ms/req at 82 coordinate<br/>0.452199ms/coordinate<br/>Radius 30m: <br/>316.653ms/req at 82 coordinate<br/>3.86162ms/coordinate | Default radius: <br/>4.53585ms/req at 82 coordinate<br/>0.0553152ms/coordinate<br/>Radius 5m: <br/>4.53168ms/req at 82 coordinate<br/>0.0552643ms/coordinate<br/>Radius 10m: <br/>15.3497ms/req at 82 coordinate<br/>0.187191ms/coordinate<br/>Radius 15m: <br/>37.2666ms/req at 82 coordinate<br/>0.454471ms/coordinate<br/>Radius 30m: <br/>317.507ms/req at 82 coordinate<br/>3.87204ms/coordinate |
| match_mld | Default radius: <br/>2.96919ms/req at 82 coordinate<br/>0.0362097ms/coordinate<br/>Radius 5m: <br/>3.03199ms/req at 82 coordinate<br/>0.0369755ms/coordinate<br/>Radius 10m: <br/>11.0508ms/req at 82 coordinate<br/>0.134766ms/coordinate<br/>Radius 15m: <br/>28.855ms/req at 82 coordinate<br/>0.35189ms/coordinate<br/>Radius 30m: <br/>339.895ms/req at 82 coordinate<br/>4.14507ms/coordinate | Default radius: <br/>2.90126ms/req at 82 coordinate<br/>0.0353813ms/coordinate<br/>Radius 5m: <br/>2.89676ms/req at 82 coordinate<br/>0.0353263ms/coordinate<br/>Radius 10m: <br/>10.5658ms/req at 82 coordinate<br/>0.128851ms/coordinate<br/>Radius 15m: <br/>27.0996ms/req at 82 coordinate<br/>0.330484ms/coordinate<br/>Radius 30m: <br/>318.225ms/req at 82 coordinate<br/>3.88079ms/coordinate |
| packedvector | random write:<br/>std::vector 10286.3 ms<br/>util::packed_vector 73749.6 ms<br/>slowdown: 7.16967<br/>random read:<br/>std::vector 8367.15 ms<br/>util::packed_vector 30493.6 ms<br/>slowdown: 3.64444 | random write:<br/>std::vector 9785.35 ms<br/>util::packed_vector 77921.4 ms<br/>slowdown: 7.96307<br/>random read:<br/>std::vector 8441.8 ms<br/>util::packed_vector 33386.6 ms<br/>slowdown: 3.95492 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>593.692ms<br/>0.593692ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>376.697ms<br/>0.376697ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>759.347ms<br/>0.759347ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>161.491ms<br/>0.161491ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>102.788ms<br/>0.102788ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>143.892ms<br/>0.143892ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>159.54ms<br/>0.15954ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>103.405ms<br/>0.103405ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>143.685ms<br/>0.143685ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>599.422ms<br/>0.599422ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>378.552ms<br/>0.378552ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>767.445ms<br/>0.767445ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>162.035ms<br/>0.162035ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>103.86ms<br/>0.10386ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>144.881ms<br/>0.144881ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>161.261ms<br/>0.161261ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>103.658ms<br/>0.103658ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>144.737ms<br/>0.144737ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>737.602ms<br/>0.737602ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>472.685ms<br/>0.472685ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>965.238ms<br/>0.965238ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>325.415ms<br/>0.325415ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>172.545ms<br/>0.172545ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>319.493ms<br/>0.319493ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>296.798ms<br/>0.296798ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>169.39ms<br/>0.16939ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>320.065ms<br/>0.320065ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>729.541ms<br/>0.729541ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>473.043ms<br/>0.473043ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>957.923ms<br/>0.957923ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>286.521ms<br/>0.286521ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>168.97ms<br/>0.16897ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>311.269ms<br/>0.311269ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>284.686ms<br/>0.284686ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>168.434ms<br/>0.168434ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>308.441ms<br/>0.308441ms/req |
| rtree | 1 result:<br/>208.211ms ->  0.0208211 ms/query<br/>10 results:<br/>243.154ms ->  0.0243154 ms/query | 1 result:<br/>207.004ms ->  0.0207004 ms/query<br/>10 results:<br/>242.165ms ->  0.0242165 ms/query |
<!-- BENCHMARK_RESULTS_END -->